move functions in the plugin extension
